+++ endpoints/_includes/functions.php
@@ -17,7 +17,7 @@ static $settings_whitelist = [
     'config' => [ 'description' => '' , 'default' => null, 'overridable' => true, 'hidden' => true],
     'custom_css' => [ 'description' => '' , 'default' => 'td { font-size: 36px; }', 'overridable' => true, 'hidden' => false],
     'custom_query' => ['description' => '', 'default' => '&sort_results_by_distance=1&long_val={LONGITUDE}&lat_val={LATITUDE}&geo_width={SETTING_MEETING_SEARCH_RADIUS}&weekdays={DAY}', 'overridable' => true, 'hidden' => false],
-    'digit_map_search_type' => [ 'description' => '', 'default' => ['1' => SearchType::VOLUNTEERS, '2' => SearchType::MEETINGS, '3' => SearchType::JFT], 'overridable' => true, 'hidden' => false],
+    'digit_map_search_type' => [ 'description' => '', 'default' => ['1' => SearchType::VOLUNTEERS, '2' => SearchType::MEETINGS, '3' => SearchType::JFT, '0' => SearchType::DIALBACK], 'overridable' => true, 'hidden' => false],
     'digit_map_location_search_method' => [ 'description' => '', 'default' => ['1' => LocationSearchMethod::VOICE, '2' => LocationSearchMethod::DTMF, '3' => SearchType::JFT], 'overridable' => true, 'hidden' => false],
     'extension_dial' => [ 'description' => '', 'default' => false, 'overridable' => true, 'hidden' => false],
     'fallback_number' => [ 'description' => '' , 'default' => '', 'overridable' => true, 'hidden' => false],
@@ -99,6 +99,7 @@ class SearchType
     const VOLUNTEERS = 1;
     const MEETINGS = 2;
     const JFT = 3;
+    const DIALBACK = 1000;
 }
 
 class LocationSearchMethod