+++ RELEASENOTES.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ### 3.5.0 (UNRELEASED)
+* Add the ability to "dialback". [#209]
 * Fix for SMS when Forced Caller ID is set. [#307]
 
 ### 3.4.0 (September 8, 2019)